+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Bump axios from 1.7.9 to 1.8.3 in /gerbera-web
 - Fix SQL for changing persistent object flag in database
 - Improve compile info output
+- Limit strings to column size
 - Remove legacy compile option SQLITE_BACKUP_ENABLED
 - Reorganise installation scripts
 - Support "natural" file sorting